requirements.txt
# Core Dependencies (already installed)
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# LangChain & LLM (already installed)
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1

# Embeddings & Vector Store (already installed)
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Data Validation (already installed)
pandera>=0.17.0

# MLOps (already installed)
mlflow>=2.9.0

# Utilities (already installed)
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pydantic>=2.0.0
loguru>=0.7.0

# API Framework (NEW - add these)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # For file uploads

# Testing (optional, see requirements-dev.txt)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# httpx>=0.25.0  # For FastAPI testing
